Handle new "active" column in raw dataframe
The raw integration dataframe now contains a column "active", which says whether the region is still in need of further refinement. In the raw data, this is an integer value; in the augmented dataframe, it is a logical value.

In addition, adding this turned up a bug in canonicalize_dim_names; it would fail unless the dimension names were the last columns in the dataframe. This is now fixed.